% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_excess.R
\name{bind_excess}
\alias{bind_excess}
\title{Bind excess communting partial tables}
\usage{
bind_excess(tabflows, tabcost, idpol, idori, iddes, idflow, iddist,
  variable, modal)
}
\arguments{
\item{tabflows}{A long table of flows}

\item{tabcost}{A squared matrix of cost}

\item{idpol}{A squared matrix of cost}

\item{idori}{A squared matrix of cost}

\item{iddes}{A squared matrix of cost}

\item{variable}{A squared matrix of cost}

\item{modal}{A squared matrix of cost}
}
\value{
A squared matrix of flows
}
\description{
This function allows you to create a distance cost matrix
}
\examples{
# Import data


}
