% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_excess_job.R
\name{bind_excess_job}
\alias{bind_excess_job}
\title{Bind excess communting partial tables}
\usage{
bind_excess_job(tabflows, matcost, idpol, idori, iddes, idflow, variable,
  modal)
}
\arguments{
\item{tabflows}{A long table of flows}

\item{matcost}{A squared matrix of cost}

\item{idpol}{A squared matrix of cost}

\item{idori}{A squared matrix of cost}

\item{iddes}{A squared matrix of cost}

\item{idflow}{A squared matrix of cost}

\item{variable}{A squared matrix of cost}

\item{modal}{A squared matrix of cost}
}
\value{
A squared matrix of flows
}
\description{
This function allows you to create a distance cost matrix
}
\examples{
# Import data


}
