% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbd_city.R
\name{cbd_city}
\alias{cbd_city}
\title{DF transform to CBD urban model}
\usage{
cbd_city(pol, id, cand, tabflows, idori, iddes, idflow)
}
\arguments{
\item{pol}{An sf object of the cities}

\item{id}{A character string of the column containing the id of the pol object}

\item{cand}{A character string of the column containing binary (1, 0) candidate value of the pol object (main city must be equal to 1 and 0 for the rest)}

\item{tabflows}{A data.frame of flows between origins and destinations (long format matrix containing, at least, origins, destinations, flows)}

\item{idori}{A character string giving the origin field name in tabflows}

\item{iddes}{A character string giving the destination field name in tabflows}

\item{idflow}{A character string giving the flow field name in tabflows}
}
\value{
A data.frame of flows as tabflows with origins and destinations changed
}
\description{
This function allows you to change the origins and destination of flows
stored into a dataframe to simulate a Central Business District urban model. City considered as
the main city of the region is designed as candidate and flows (destinations only) are moved
from non-candidate cities to the candidate city. The origins are moved into from the candidate city to the non-candidate city,
so that every jobs are in the main city, and workers in the suburbs.
}
