% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_com.R
\name{coord_com}
\alias{coord_com}
\title{cities coordinates}
\usage{
coord_com(pol)
}
\arguments{
\item{pol}{An sf object of the cities}
}
\value{
A data.frame with the same variables as pol, centroïds (lat,lon) and no geometries
}
\description{
This function allows you to store the coordinates of each cities in a dataframe
}
\examples{
# Import data
data(pol)

coordCom <- coord_com(pol)

coordCom[1:10,]

}
