% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_matrix.R
\name{cost_matrix}
\alias{cost_matrix}
\title{Create square distance matrix}
\usage{
cost_matrix(tabcost, idpol, idori, iddes, iddist)
}
\arguments{
\item{tabcost}{A data.frame of flows between origins and destinations (long format matrix containing, at least, 3 column : origins, destinations, flows)}

\item{idpol}{A character vector id spatial units}

\item{idori}{A character string giving the origin field name in tabflows}

\item{iddes}{A character string giving the destination field name in tabflows}

\item{iddist}{A character string giving the flow field name in tabflows}
}
\value{
A squared Origins (rows) and Destination (columns) matrix of flows
}
\description{
This function allows you to create an Origin/Destination squared matrix from a Origin/Destination long format matrix
}
