% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tabflows.R
\name{create_tabflows}
\alias{create_tabflows}
\title{Create an O/D long table}
\usage{
create_tabflows(tabindiv, idori, iddes, varwgt = NULL, variable = NULL,
  label = NULL)
}
\arguments{
\item{tabindiv}{A data.frame of individuals between origins and destinations (long format matrix containing, at least, origins, destinations for each individuals)}

\item{idori}{A character string giving the origin field name in tabindiv}

\item{iddes}{A character string giving the destination field name in tabindiv}

\item{varwgt}{Default to NULL ; a character string giving the weight field name in tabindiv}

\item{variable}{Default to NULL ; a character string giving the name of the field in tabindiv in wich selected label will be filtered}

\item{label}{Default to NULL ; a character string giving the value of the variable to be filtered (or keeped)}
}
\value{
A data.frame of three column : ORI, DES, WGT
}
\description{
This function allows you to create an Origin/Destination matrix (in a long table format) from a table of individuals.
It may deal with weighted individuals and filter these individuals
}
\examples{
# Import data
data(tabindiv)
idori <- "ORI"
iddes <- "DES"

tabflows <- create_tabflows(tabindiv, idori, iddes)

tabflows[1:10,]

# we will now weight individuals with the column "WGT"
varwgt <- "WGT"

tabflows <- create_tabflows(tabindiv, idori, iddes, varwgt)

tabflows[1:10,]

# now we will only extract the values where the SCP is equal to "3"
variable <- "SCP"
label <- "3"

tabflowsW <- create_tabflows(tabindiv, idori, iddes, varwgt, variable, label)

tabflowsW[1:10,]

}
