% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excess_commuting.R
\name{excess_commuting}
\alias{excess_commuting}
\title{Excess commuting}
\usage{
excess_commuting(matflows, matcost)
}
\arguments{
\item{matflows}{A squared matrix of flows}

\item{matcost}{A squared matrix of cost}
}
\value{
A squared matrix of flows
}
\description{
This function allows you to create a distance cost matrix
}
