% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finger_plan.R
\name{finger_plan}
\alias{finger_plan}
\title{DF transform to finger plan urban model}
\usage{
finger_plan(pol, idpol, cand, tabflows, idori, iddes, idflow)
}
\arguments{
\item{pol}{An sf object of the cities}

\item{idpol}{A character string of the column containing the id of the pol object}

\item{cand}{A character string of the column containing binary (1, 0) candidate value of the pol object (1 must be equal to city containing a railroad station)}

\item{tabflows}{A data.frame of flows between origins and destinations (long format matrix containing, at least, origins, destinations, flows)}

\item{idori}{A character string giving the origin field name in tabflows}

\item{iddes}{A character string giving the destination field name in tabflows}

\item{idflow}{A character string giving the flow field name in tabflows}
}
\value{
A data.frame of flows as tabflows with origins and destinations changed
}
\description{
This function allows you to change the origins and destination of flows
stored into a dataframe to simulate a finger-plan urban model. Cities containing
railroad station are designed as candidate and flows (origins and destinations)
are moved from non-candidate cities to the nearest candidate cities (using osm network)
}
\examples{
# Import data

idpol <- "idpol"
cand <- "cand"
data(polL93)
data(tabflows)
idori <- "ORI"
iddes <- "DES"
idflow <- "FLOW"

polL93$cand <- sample(0:1, size = nrow(polL93), replace = TRUE)

fg_flows <- finger_plan(
pol = polL93,
idpol = idpol,
cand = cand,
tabflows = tabflows,
idori = idori,
iddes = iddes,
idflow = idflow)

fg_flows[1:10,]

}
