% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icdr.R, R/label_icdr.R
\name{label_icdr}
\alias{label_icdr}
\title{Label flows as ICDR}
\usage{
label_icdr(tabflows, idori, iddes, idflow, pol, idpol)

label_icdr(tabflows, idori, iddes, idflow, pol, idpol)
}
\arguments{
\item{tabflows}{A data.frame of flows between origins and destinations (long format matrix containing, at least, origins, destinations, flows)}

\item{idori}{A character string giving the origin field name in tabflows}

\item{iddes}{A character string giving the destination field name in tabflows}

\item{idflow}{A character string giving the flow field name in tabflows}

\item{pol}{An sf object of the cities}

\item{idpol}{A character string identifier of cities}

\item{tabflows}{A data.frame of flows between origins and destinations (long format matrix containing, at least, origins, destinations, flows)}

\item{idori}{A character string giving the origin field name in tabflows}

\item{iddes}{A character string giving the destination field name in tabflows}

\item{idflow}{A character string giving the flow field name in tabflows}
}
\value{
A data.frame of flows as tabflows with ICDR labels

A data.frame of flows as tabflows with ICDR labels
}
\description{
This function bla bla

This function bla bla
}
