% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapvalues_.R
\name{mapvalues_}
\alias{mapvalues_}
\title{Replace specified values with new values, in a vector or factor (from dplyr)}
\usage{
mapvalues_(x, from, to, warn_missing = TRUE)
}
\arguments{
\item{x}{the factor or vector to modify}

\item{from}{a vector of the items to replace}

\item{to}{a vector of replacement values}

\item{warn_missing}{print a message if any of the old values are not actually present in x}
}
\description{
Item in x that match items from will be replaced by items in to, matched by position. For example, items in x that match the first element in from will be replaced by the first element of to.
}
\details{
If x is a factor, the matching levels of the factor will be replaced with the new values.
The related revalue function works only on character vectors and factors, but this function works on vectors of any type and factors.
}
\examples{
x <- c("a", "b", "c")
mapvalues_(x, c("a", "c"), c("A", "C"))

# Works on factors
y <- factor(c("a", "b", "c", "a"))
mapvalues_(y, c("a", "c"), c("A", "C"))

# Works on numeric vectors
z <- c(1, 4, 5, 9)
mapvalues_(z, from = c(1, 5, 9), to = c(10, 50, 90))

}
