% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nystuen_dacey.R
\name{nystuen_dacey}
\alias{nystuen_dacey}
\title{Dominant flows}
\usage{
nystuen_dacey(tabflows, idori, iddes, idflow, weight, threspct, pol, idpol)
}
\arguments{
\item{tabflows}{A data.frame of flows between origins and destinations (long format matrix containing, at least, 3 column : origins, destinations, flows)}

\item{idori}{A character string giving the origin field name in tabflows}

\item{iddes}{A character string giving the destination field name in tabflows}

\item{idflow}{A character string giving the flow field name in tabflows}

\item{weight}{A character string, chosen between, "destination", "origin" or "internal" to weigths the flows}

\item{threspct}{A threshold (see 'Details')}

\item{pol}{An sf object of the cities}

\item{idpol}{A character string identifier of cities

This function selects which flow (fij or fji) must be kept. If the ratio weight of destination (wj) / weight of origin (wi)
is greater than the treshold, then fij is selected and fji is not. This function can perform the second criterion of the Nystuen & Dacey's dominants flows analysis.
As the output is a boolean matrix, use element-wise multiplication to get flows intensity.}
}
\value{
A boolean matrix of selected flows
}
\description{
This function selects the flows to be keeped in a large matrix of flows responding to the Nystuen & Dacey's dominants flows criterion.
}
\examples{
# Import data
data(tabflows)
idori <- "ORI"
iddes <- "DES"
idflow <- "FLOW"
weight <- "destination"
threspct <- 3
data(pol)
idpol <- "idpol"

domflow <- nystuen_dacey(
tabflows,
idori,
iddes,
idflow,
weight,
threspct,
pol,
idpol)

domflow[1]
domflow[2]

}
