% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pol.R
\docType{data}
\name{pol}
\alias{pol}
\title{polygons}
\description{
Polygon data of cities in Paris region (stored as sf)
Fields: \cr
\itemize{
\item{name}
\item{idpol}
}
}
\examples{
## pol
data(pol)
str(pol)
}
\references{
\url{http://data.iledefrance.fr}
}
