% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pol_aggregate(tidyverse).R
\name{pol_aggregate}
\alias{pol_aggregate}
\title{Aggregate cities of a sf object}
\usage{
pol_aggregate(before, after, pol, idpol, namepol, nameAgr)
}
\arguments{
\item{before}{A list of the identifiers to be replaced}

\item{after}{A character string of the identifier replacement}

\item{pol}{An sf object of the cities}

\item{idpol}{A character string of the column containing the id of the pol object}

\item{namepol}{A character string of the column containing the names of the pol object}

\item{nameAgr}{A character string of the name to be changed for the aggregated city}
}
\value{
An sf object with a new id of aggregate city, and the polygons aggregated
}
\description{
This function allows you to aggregate cities of a spatial dataframe from designed identifier
}
\examples{
# Import data
before <-  c(75101, 75102, 75103,75104, 75105,75106,75107, 75108, 75109,
             75110,75111, 75112, 75113, 75114,75115, 75116, 75117,75118, 75119, 75120)
after <- 75056
data(pol)
idpol <- "idpol"
namepol <- "name"
nameAgr <- "paris"

polAggregate <- pol_aggregate(before, after, pol, idpol, namepol, nameAgr)

plot(polAggregate)

}
