% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polL93.R
\docType{data}
\name{polL93}
\alias{polL93}
\title{polL93}
\description{
Polygon data of cities in Paris region(stored as sf in "lambert 93" projection)
Fields: \cr
\itemize{
\item{nomcom}
\item{numdep}
\item{insee}
\item{shape_leng}
\item{shape_area}
}
}
\examples{
## polL93
data(polL93)
str(polL93)
}
\references{
\url{http://data.iledefrance.fr}
}
