% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_tab.R
\name{pop_tab}
\alias{pop_tab}
\title{Total flows of a DF}
\usage{
pop_tab(tabflows, idori, iddes, idflow, iddist)
}
\arguments{
\item{tabflows}{A data.frame of flows between origins and destinations (long format matrix containing, at least, origins, destinations, flows)}

\item{idori}{identifiant ori}

\item{iddes}{identifiant des}

\item{idflow}{identifiant flux}

\item{iddist}{identifiant dist}
}
\value{
A data.frame of totals origins, destinations, internals flows and mean of distance for origins and destination of each city
}
\description{
This function allows you to store the totals of origins, destinations and intrenals flows for each city in a dataframe,
from a long format matrix of flows.
}
\examples{
data(tabflows)

popTab <- pop_tab(tabflows = tabflows, idori = "ORI",
iddes = "DES", idflow = "FLOW", iddist = "DIST")

popTab[10:10,]

}
