% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_matflows.R
\name{prepare_matflows}
\alias{prepare_matflows}
\title{O/D Matrix from individuals}
\usage{
prepare_matflows(tabindiv, idori, iddes, varwgt = NULL,
  variable = NULL, label = NULL)
}
\arguments{
\item{tabindiv}{A data.frame of individuals between origins and destinations (long format matrix containing, at least, origins, destinations for each individuals)}

\item{idori}{A character string giving the origin field name in tabindiv}

\item{iddes}{A character string giving the destination field name in tabindiv}

\item{varwgt}{Default to NULL ; a character string giving the weight field name in tabindiv}

\item{variable}{Default to NULL ; a character string giving the name of the field in tabindiv in wich selected label will be filtered}

\item{label}{Default to NULL ; a character string giving the value of the variable to be filtered (or keeped)}
}
\value{
A squared matrix of flows
}
\description{
This function allows you to create an O/D matrix from a table of individuals.
It may deal with weighted individuals and filter these individuals
}
\examples{
# Import data
data(tabindiv)
idori <- "ORI"
iddes <- "DES"


matFlows <- prepare_matflows(tabindiv, idori, iddes)

matFlows[1:10,]

# we will now weight individuals with the column "WGT"
varwgt <- "WGT"

matFlows <- prepare_matflows(tabindiv, idori, iddes, varwgt)

matFlows[1:10,]

# now we will only extract the values where the SCP is equal to "3"
variable <- "SCP"
label <- "3"

matFlowsW <- prepare_matflows(tabindiv, idori, iddes, varwgt, variable, label)

matFlowsW[1:10,]

}
