% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relocate_one.R
\name{relocate_one}
\alias{relocate_one}
\title{Create a dictionnary of id change}
\usage{
relocate_one(pol, idpol, cand)
}
\arguments{
\item{pol}{An sf object of the cities}

\item{idpol}{A character string of the column containing the id of the pol object}

\item{cand}{A character string of the column containing binary (1, 0) candidate value of the pol object}
}
\value{
A data.frame with two column : OLD (containing the id to be changed) and NEW (containing the new id)
}
\description{
This function is a low-level one and used in each *_city function (cbd_city, tod_city, finger_city ...etc.).
this gives a dictionary which indicates by which new identifier the old one must be changed, the latter
answering the criterion of proximity and to the candidacy of the city
}
\examples{
# Import data
data(polL93)
idpol <- "idpol"
cand <- "cand"

polL93$cand <- sample(0:1, size = nrow(polL93), replace = TRUE)

dictioTransfer <- relocate_one(polL93, idpol, cand)

dictioTransfer[1:10,]

}
