% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/road.R
\docType{data}
\name{road}
\alias{road}
\title{road linestring}
\description{
linestring of road in Paris region (stored as sf)
Fields: \cr
\itemize{
\item{fclass}
\item{name}
\item{oneway}
\item{maxspeed}
}
}
\examples{
## road
data(road)
str(road)
}
\references{
\url{http://download.geofabrik.de/}
}
