% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routing_machine.R
\name{routing_machine}
\alias{routing_machine}
\title{Nearest neighbours using road network}
\usage{
routing_machine(road, pol, idpol)
}
\arguments{
\item{road}{An sf object of street network}

\item{pol}{An sf object of the cities}

\item{idpol}{A character string identifier of cities}
}
\value{
a squared matrix of distance in meters
}
\description{
This function allows you to find the nearest neighbours from polygons centroïd to one another using the road network.
It creates a squared distance matrix between every city in meters.
}
\examples{
# Import data
data(road)
data(pol)
idpol <- "idpol"

matDist <- routing_machine(road,pol,idpol)

matDist[1:10,]

}
