% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabflows.R
\docType{data}
\name{tabflows}
\alias{tabflows}
\title{Commuters}
\description{
Data on commuters between Urban Areas in Paris region in 2014.
Fields: \cr
\itemize{
\item{ORI: Code of the urban area of residence}
\item{DES: Code of the urban area of work}
\item{FLOW: Number of commuters between i and j}
\item{DIST: distance between the urban area of work and the urban area of residence}
}
}
\examples{
## tabflows
data(tabflows)
}
