% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabindiv.R
\docType{data}
\name{tabindiv}
\alias{tabindiv}
\title{tabindiv}
\description{
dataframe of individuals and their daily travel to work inside the Paris urban region(stored as df)
Fields: \cr
\itemize{
\item{ORI}
\item{DES}
\item{SCP}
\item{MODE}
\item{WGT}
}
}
\examples{
## tabindiv
data(tabindiv)
str(tabindiv)
}
\references{
\url{https://www.insee.fr/fr/statistiques/2862065}
}
