% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_tab.R
\name{pop_tab}
\alias{pop_tab}
\title{Total flows of a DF}
\usage{
pop_tab(tabflows, tabdist, idori, iddes, idflow, iddist)
}
\arguments{
\item{tabflows}{A data.frame of flows (long format: origin, destination, flow)}

\item{tabdist}{asf}

\item{idori}{identifiant ori}

\item{iddes}{identifiant des}

\item{idflow}{identifiant flux}

\item{iddist}{identifiant dist}
}
\value{
A data.frame of totals origins, destinations, internals flows and sum of distance for origins and destination of each city
}
\description{
This function computes the margins of the OD matrix: emissions, attractions, intraflows for each spatial unit
from a long format matrix of flows.
}
\examples{
data(tabflows)


}
